// Frontend-angular/src/app/models/user.model.ts
export class User {

 

  id: number;
  userName: string;
  password: string;
  firstName: string;
  lastName: string;
  email: string;
  contactNumber: number;
  regCode: string;
  role: string;
  linkedinUrl: string;
  yearsofExperience: number;
  active: boolean;
  confirmedSignup: boolean;
  resetPassword: boolean;
  resetPasswordDate: Date;
  
}